requirements.txt
polars==1.17.1
pyyaml
python-dotenv
pydantic
